+++ src/main/java/org/truffleruby/parser/YARPMultiWriteNodeTranslator.java
@@ -20,6 +20,9 @@ import org.truffleruby.core.cast.SplatCastNode;
 import org.truffleruby.core.cast.SplatCastNodeGen;
 import org.truffleruby.language.RubyNode;
 
+import java.util.ArrayList;
+import java.util.List;
+
 /** Translate Nodes.MultiWriteNode node.
  *
  * NOTE: cannot inherit from YARPBaseTranslator because it returns AssignableNode instead of RubyNode. */
@@ -28,6 +31,10 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
     private final Nodes.MultiWriteNode node;
     private final RubyLanguage language;
     private final YARPTranslator yarpTranslator;
+    /** Nodes to initialize local variables before assign values. Store in the local variables receivers of attribute
+     * (a.b = ...) and reference (a[b] = ...) assignments as well as fully qualified constants' modules (A::B = ...).
+     * They should be evaluated before values to conform with the left-to-right order semantic. */
+    private final List<RubyNode> prolog;
 
     public YARPMultiWriteNodeTranslator(
             Nodes.MultiWriteNode node,
@@ -36,6 +43,7 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
         this.node = node;
         this.language = language;
         this.yarpTranslator = yarpTranslator;
+        this.prolog = new ArrayList<>();
     }
 
     public RubyNode translate() {
@@ -70,12 +78,20 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
             postNodes[i] = node.rights[i].accept(this);
         }
 
-        rubyNode = new MultipleAssignmentNode(
+        RubyNode multipleAssignmentNode = new MultipleAssignmentNode(
                 preNodes,
                 restNode,
                 postNodes,
                 splatCastNode,
                 rhsNode);
+
+        if (prolog.isEmpty()) {
+            rubyNode = multipleAssignmentNode;
+        } else {
+            RubyNode prologSequence = YARPBaseTranslator.sequence(prolog.toArray(RubyNode.EMPTY_ARRAY));
+            rubyNode = YARPBaseTranslator.sequence(prologSequence, multipleAssignmentNode);
+        }
+
         return rubyNode;
     }
 
@@ -87,12 +103,30 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
 
     @Override
     public AssignableNode visitCallTargetNode(Nodes.CallTargetNode node) {
+        // store receiver in a local variable to evaluate before assigned value
+        var receiverExpression = new YARPExecutedOnceExpression("receiver", node.receiver, yarpTranslator);
+        RubyNode writeReceiverNode = receiverExpression.getWriteNode();
+        Nodes.Node readReceiver = receiverExpression.getReadYARPNode();
+
+        prolog.add(writeReceiverNode);
+        node = new Nodes.CallTargetNode(node.flags, readReceiver, node.name, node.startOffset, node.length);
+
         final RubyNode rubyNode = node.accept(yarpTranslator);
         return ((AssignableNode) rubyNode).toAssignableNode();
     }
 
     @Override
     public AssignableNode visitConstantPathTargetNode(Nodes.ConstantPathTargetNode node) {
+        // store parent lexical scope (e.g foo in foo::C = ...) in a local variable to evaluate before assigned value
+        if (node.parent != null) {
+            var parentExpression = new YARPExecutedOnceExpression("parent", node.parent, yarpTranslator);
+            RubyNode writeParentNode = parentExpression.getWriteNode();
+            Nodes.Node readParent = parentExpression.getReadYARPNode();
+
+            prolog.add(writeParentNode);
+            node = new Nodes.ConstantPathTargetNode(readParent, node.child, node.startOffset, node.length);
+        }
+
         final RubyNode rubyNode = node.accept(yarpTranslator);
         return ((AssignableNode) rubyNode).toAssignableNode();
     }
@@ -116,6 +150,33 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
 
     @Override
     public AssignableNode visitIndexTargetNode(Nodes.IndexTargetNode node) {
+        // store receiver in a local variable to evaluate before assigned value
+        var receiverExpression = new YARPExecutedOnceExpression("receiver", node.receiver, yarpTranslator);
+        RubyNode writeReceiverNode = receiverExpression.getWriteNode();
+        Nodes.Node readReceiver = receiverExpression.getReadYARPNode();
+
+        prolog.add(writeReceiverNode);
+
+        // store arguments in local variables to evaluate after receiver but before assigned values
+        final Nodes.ArgumentsNode arguments;
+        if (node.arguments != null) {
+            var argumentsReads = new Nodes.Node[node.arguments.arguments.length];
+
+            for (int i = 0; i < node.arguments.arguments.length; i++) {
+                var e = new YARPExecutedOnceExpression("argument", node.arguments.arguments[i], yarpTranslator);
+                prolog.add(e.getWriteNode());
+                argumentsReads[i] = e.getReadYARPNode();
+            }
+
+            arguments = new Nodes.ArgumentsNode(node.arguments.flags, argumentsReads, node.arguments.startOffset,
+                    node.arguments.length);
+        } else {
+            arguments = null;
+        }
+
+        node = new Nodes.IndexTargetNode(node.flags, readReceiver, arguments, node.block, node.startOffset,
+                node.length);
+
         final RubyNode rubyNode = node.accept(yarpTranslator);
         return ((AssignableNode) rubyNode).toAssignableNode();
     }
@@ -136,6 +197,8 @@ public final class YARPMultiWriteNodeTranslator extends AbstractNodeVisitor<Assi
     public AssignableNode visitMultiTargetNode(Nodes.MultiTargetNode node) {
         final var translator = new YARPMultiTargetNodeTranslator(node, language, yarpTranslator, null);
         final MultipleAssignmentNode multipleAssignmentNode = translator.translate();
+
+        prolog.addAll(translator.prolog);
 
         return multipleAssignmentNode.toAssignableNode();
     }